cleaning(core): move utils/endpoints to prismic/endpoints
