+++ packages/core/src/utils/index.ts
@@ -1,5 +1,4 @@
 export * as Cookie from "./cookie";
-export * as Endpoints from "./endpoints";
 export * as Poll from "./poll";
 export * as lib from "./lib";
 export * as str from "./str";
