+++ packages/slice-machine/components/LoginModal/index.tsx
@@ -12,7 +12,7 @@ import {
 } from "theme-ui";
 import SliceMachineModal from "@components/SliceMachineModal";
 import { checkAuthStatus, startAuth } from "@src/apiClient";
-import { buildEndpoints } from "@slicemachine/core/build/utils/endpoints";
+import { buildEndpoints } from "@slicemachine/core/build/prismic/endpoints";
 import { startPolling } from "@slicemachine/core/build/utils/poll";
 import { CheckAuthStatusResponse } from "@models/common/Auth";
 import { useSelector } from "react-redux";
